# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaText API
  description: API semplice per la piattaforma di messaggistica WASAText
  version: 1.0.0
...
security:
  - bearerAuth: []
paths:
...
  /users/{userId}/conversations/{conversationId}/messages/{messageId}/replyMessage:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
    post:
      tags:
        - message
      summary: Rispondi a un messaggio
      description: Crea un nuovo messaggio in risposta a un messaggio esistente.
      operationId: replyMessage
      requestBody:
        description: Dettagli messaggio di risposta
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  oneOf:
                    - $ref: '#/components/schemas/Text'
                    - $ref: '#/components/schemas/Photo'
        required: true
      responses:
        "201":
          description: Risposta inviata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          description: Dati non validi
        '404':
          description: Messaggio o conversazione non trovati
        '500':
          description: Errore server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Message:
      title: Message
      description: "Rappresenta un singolo messaggio."
      type: object
      properties:
        id:
          description: Identificatore unico del messaggio.
          type: integer
          example: 1
          readOnly: true
        timestamp: 
          type: string
          format: date-time
          description: Timestamp di invio del messaggio.
          example: "2023-11-05T14:30:00Z"
        content:
          description: Contenuto del messaggio.
          oneOf:
            - $ref: '#/components/schemas/Text'
            - $ref: '#/components/schemas/Photo'
        sender:
          $ref: '#/components/schemas/User'
        status:
          type: string
          enum: ["received", "read"]
          description: Stato del messaggio
          example: "read"
        commments:
          type: array
          minItems: 0
          maxItems: 500
          description: Lista di commenti sul messaggio
          items:
            $ref: '#/components/schemas/Comment'
        type:
          description: Tipo di messaggio
          type: string
          enum: ["standard", "forward", "reply"]
          example: "standard"
        replyToMessageId:
          type: integer
          description: ID del messaggio a cui si sta rispondendo.
          example: 12
    Comment:
      title: Comment
      description: "Rappresenta un singolo commento."
      type: object
      properties:
        commentId:
          description: Identificatore unico del commento.
          type: integer
          example: 1
          readOnly: true
        sender:
          $ref: '#/components/schemas/User'
        content:
          type: string
          description: Il commento come emoji
          example: "👍"
          pattern: '^.*?$'
          minLength: 1
          maxLength: 2
...
    User:
      title: User
      description: "Rappresenta un singolo utente."
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/UserId'
        username:
          $ref: '#/components/schemas/Username'
    Photo:
      description: Dati immagine codificati in Base64
      type: string
      format: byte
      pattern: "^[01]+$"
      minLength: 0
      maxLength: 9999999999999999999
    Username:
      description: Username dell'utente.
      type: string
      example: "user123"
      minLength: 3
      maxLength: 16
      pattern: "^.*?$"
    UserId:
      description: Identificatore unico dell'utente
      type: integer
      example: 1
      readOnly: true
...
    Text:
      type: string
      description: Contenuto testuale del messaggio
      example: "Ciao, come stai?"
      minLength: 1
      maxLength: 500
      pattern: '^.*?$'
  parameters:
    userId:
      schema:
        $ref: '#/components/schemas/UserId'
      name: userId
      in: path
      required: true
      description: ID utente
    conversationId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Identificatore unico della conversazione."
      name: conversationId
      in: path
      required: true
      description: ID conversazione
    messageId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Identificatore unico del messaggio."
      name: messageId
      in: path
      required: true
      description: ID messaggio
...